+++ .changeset/fc107472/changes.json
@@ -1,0 +1,49 @@
+{
+  "releases": [
+    { "name": "@voussoir/adapter-mongoose", "type": "major" },
+    { "name": "@voussoir/core", "type": "major" },
+    { "name": "@voussoir/fields", "type": "patch" }
+  ],
+  "dependents": [
+    {
+      "name": "@voussoir/test-utils",
+      "type": "patch",
+      "dependencies": ["@voussoir/adapter-mongoose", "@voussoir/core"]
+    },
+    {
+      "name": "@voussoir/cypress-project-access-control",
+      "type": "patch",
+      "dependencies": [
+        "@voussoir/test-utils",
+        "@voussoir/adapter-mongoose",
+        "@voussoir/core",
+        "@voussoir/fields"
+      ]
+    },
+    {
+      "name": "@voussoir/cypress-project-basic",
+      "type": "patch",
+      "dependencies": [
+        "@voussoir/test-utils",
+        "@voussoir/adapter-mongoose",
+        "@voussoir/core",
+        "@voussoir/fields"
+      ]
+    },
+    {
+      "name": "@voussoir/cypress-project-login",
+      "type": "patch",
+      "dependencies": [
+        "@voussoir/test-utils",
+        "@voussoir/adapter-mongoose",
+        "@voussoir/core",
+        "@voussoir/fields"
+      ]
+    },
+    {
+      "name": "@voussoir/cypress-project-twitter-login",
+      "type": "patch",
+      "dependencies": ["@voussoir/adapter-mongoose", "@voussoir/core", "@voussoir/fields"]
+    }
+  ]
+}
