+++ .changeset/fc107472/changes.md
@@ -1,0 +1,1 @@
+- Add a postConnect method to list adapters to capture all the work which needs to be done after the database has been connected to
