+++ packages/core/package.json
@@ -18,6 +18,7 @@
     "graphql-tools": "^4.0.3",
     "graphql-type-json": "^0.2.1",
     "oauth-libre": "^0.9.17",
+    "p-settle": "^2.1.0",
     "p-waterfall": "^1.0.0",
     "passport": "^0.4.0",
     "passport-twitter": "^1.0.4",
